requirements.txt
discord.py[voice]>=1.5.1